athon/requirements.txt
Django==1.7.4
Pillow==2.5.1
PyYAML==3.11
dict.sorted==1.0.0
diff-match-patch==20121119
dj-database-url==0.3.0
dj-static==0.0.6
django-autocomplete-light==2.0.7
django-enumfield==1.1.2
django-filter==0.9.2
django-rest-swagger==0.2.8
django-tagging==0.3.4
django-taggit==0.12.1
django-uuid-upload-path==1.0.0
djangorestframework==2.3.14
drf-toolbox==0.1.7
imagestore==2.9.1
psycopg2==2.5.4
py==1.4.26
pytest==2.6.4
pytest-django==2.8.0
pytz==2014.10
six==1.9.0
sorl-thumbnail==12.2
static3==0.5.1
swapper==0.2.1
uWSGI==2.0.6
wsgiref==0.1.2
